# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "description": "API для работы с Task manager",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить список всех пользователей",
                "operationId": "d7a57f4d7de18fb0f92501172472c9b6",
                "responses": {
                    "200": {
                        "description": "Успешно получен список пользователей",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "ID пользователя",
                                                "type": "integer"
                                            },
                                            "first_name": {
                                                "description": "Имя пользователя",
                                                "type": "string"
                                            },
                                            "last_name": {
                                                "description": "Фамилия пользователя",
                                                "type": "string"
                                            },
                                            "username": {
                                                "description": "Имя пользователя для входа",
                                                "type": "string"
                                            },
                                            "email": {
                                                "description": "Электронная почта пользователя",
                                                "type": "string"
                                            },
                                            "condition_id": {
                                                "description": "ID состояния пользователя",
                                                "type": "integer"
                                            },
                                            "created_at": {
                                                "description": "Дата создания пользователя",
                                                "type": "string",
                                                "format": "date-time"
                                            },
                                            "updated_at": {
                                                "description": "Дата последнего обновления пользователя",
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Создать нового пользователя",
                "operationId": "9fea6f2adcb69614c521a7733358382d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "Имя пользователя",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Фамилия пользователя",
                                        "type": "string"
                                    },
                                    "username": {
                                        "description": "Имя пользователя для входа",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Электронная почта пользователя",
                                        "type": "string"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния пользователя",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пользователь успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "first_name": {
                                            "description": "Имя пользователя",
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "description": "Фамилия пользователя",
                                            "type": "string"
                                        },
                                        "username": {
                                            "description": "Имя пользователя для входа",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Электронная почта пользователя",
                                            "type": "string"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния пользователя",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/users/{id}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Получить информацию о пользователе",
                "operationId": "40a5104e23b52137e78c3d3ac48b88a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно получена информация о пользователе",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "first_name": {
                                            "description": "Имя пользователя",
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "description": "Фамилия пользователя",
                                            "type": "string"
                                        },
                                        "username": {
                                            "description": "Имя пользователя для входа",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Электронная почта пользователя",
                                            "type": "string"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния пользователя",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Обновить данные пользователя",
                "operationId": "b166772cec9714067040e77cf14beeba",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "description": "Имя пользователя",
                                        "type": "string"
                                    },
                                    "last_name": {
                                        "description": "Фамилия пользователя",
                                        "type": "string"
                                    },
                                    "username": {
                                        "description": "Имя пользователя для входа",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Электронная почта пользователя",
                                        "type": "string"
                                    },
                                    "condition_id": {
                                        "description": "ID состояния пользователя",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пользователь успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "ID пользователя",
                                            "type": "integer"
                                        },
                                        "first_name": {
                                            "description": "Имя пользователя",
                                            "type": "string"
                                        },
                                        "last_name": {
                                            "description": "Фамилия пользователя",
                                            "type": "string"
                                        },
                                        "username": {
                                            "description": "Имя пользователя для входа",
                                            "type": "string"
                                        },
                                        "email": {
                                            "description": "Электронная почта пользователя",
                                            "type": "string"
                                        },
                                        "condition_id": {
                                            "description": "ID состояния пользователя",
                                            "type": "integer"
                                        },
                                        "created_at": {
                                            "description": "Дата создания пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        },
                                        "updated_at": {
                                            "description": "Дата последнего обновления пользователя",
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Удалить пользователя",
                "operationId": "ce06860a51ab016ef66e753544fe4563",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь успешно удален"
                    }
                }
            }
        }
    },
...
}